Added python-cinderclient dependancy + cleanup
python-cinderclient was required, but missing from requirements.txt.
Thanks to @tourunen for reporting.

Added help() and debug() methods. Help displays the usage text, but
inside bpython. debug() enables python logging to the console which can
be useful for debugging connection issues or understanding what REST
api the python code is actually using (it would be nice if one never
need to do this!).

Cleaned up the code a bit.